requirements.txt
Django==3.2.6
djangorestframework==3.12.4
elasticsearch-dsl==7.4.0
django_elasticsearch_dsl
django_elasticsearch_dsl_drf
django-ckeditor
psycopg2-binary==2.9.6
pillow